refactor: 라우터 로직 개선
